リスト登録ダイアログを途中まで作った
リストTLのIDキャストを消した

+++ Yukari/src/main/java/shibafu/yukari/fragment/tabcontent/DefaultTweetListFragment.java
@@ -345,7 +345,7 @@ public class DefaultTweetListFragment extends TweetListFragment implements Statu
                         responseList = twitter.getFavorites(targetUser.getId(), paging);
                         break;
                     case TabType.TABTYPE_LIST:
-                        responseList = twitter.getUserListStatuses((int)listId, paging);
+                        responseList = twitter.getUserListStatuses(listId, paging);
                         break;
                 }
                 if (!params[0].isSaveLastPaging()) {
